Add THNK Rooms (Relay) adapter (#38)
THNK Rooms (or THNK Relay) is a new adapter that allows to connect using a simple rooms system. Any user can create a room with any name, and any other user can connect to the room with the same name.

It will be extended in the future to allow querying room names based on numerous criterion, to allow doing matchmaking.